fix(report_sonar): fix some PR comments
