+++ tools/devsecops_engine_tools/engine_core/src/domain/model/gateway/vulnerability_management_gateway.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta, abstractmethod
 
 from devsecops_engine_tools.engine_core.src.domain.model.vulnerability_management import VulnerabilityManagement
-
+from devsecops_engine_tools.engine_core.src.domain.model.gateway.devops_platform_gateway import DevopsPlatformGateway
 
 class VulnerabilityManagementGateway(metaclass=ABCMeta):
     @abstractmethod
@@ -27,3 +27,17 @@ class VulnerabilityManagementGateway(metaclass=ABCMeta):
         self, engagement_name, dict_args, secret_tool, config_tool
     ):
         "get_active_engagements"
+
+    @abstractmethod
+    def send_report(
+        self,
+        compact_remote_config_url: str,
+        source_code_management_uri: str,
+        environment: str,
+        token_vulnerability_management: str,
+        token_cmdb: str,
+        config_tool: any,
+        devops_platform_gateway: DevopsPlatformGateway,
+        project_key: str
+    ):
+        "send_report"
