--- tools/devsecops_engine_tools/engine_utilities/sonarqube/domain/model/gateways/vulnerability_management_gateway.py
@@ -1,22 +1,0 @@
-from abc import (
-    ABCMeta,
-    abstractmethod
-)
-from devsecops_engine_tools.engine_core.src.domain.model.gateway.devops_platform_gateway import (
-    DevopsPlatformGateway,
-)
-
-class VulnerabilityManagementGateway(metaclass=ABCMeta):
-    @abstractmethod
-    def send_report(
-        self,
-        compact_remote_config_url: str,
-        source_code_management_uri: str,
-        environment: str,
-        token_vulnerability_management: str,
-        token_cmdb: str,
-        config_tool: any,
-        devops_platform_gateway: DevopsPlatformGateway,
-        project_key: str
-    ):
-        "send_report"
