+++ tools/devsecops_engine_tools/engine_utilities/sonarqube/infrastructure/helpers/utils.py
@@ -6,9 +6,3 @@ def set_repository(pipeline_name, source_code_management):
     else:
         splittedPipeline = pipeline_name.split('_MR_')
         return source_code_management + '?path=/' + splittedPipeline[1]
-
-def set_environment(branch):
-    if branch in ["trunk", "master"]:
-        return "Production"
-    else:
-        return "Development"
